adding save method to serve controller
