+++ Gateway/crds-angular.test/controllers/ServeControllerTest.cs
@@ -5,6 +5,7 @@ using System.Web.Http;
 using System.Web.Http.Controllers;
 using System.Web.Http.Results;
 using crds_angular.Controllers.API;
+using crds_angular.Models.Crossroads;
 using crds_angular.Models.Crossroads.Serve;
 using crds_angular.Services.Interfaces;
 using MinistryPlatform.Translation.Services.Interfaces;
@@ -40,12 +41,6 @@ namespace crds_angular.test.controllers
             _fixture.Request = new HttpRequestMessage();
             _fixture.Request.Headers.Authorization = new AuthenticationHeaderValue(_authType, _authToken);
             _fixture.RequestContext = new HttpRequestContext();
-        }
-
-        [Test]
-        public void SaveFamilySignUp()
-        {
-            
         }
 
         [Test]
