+++ Gateway/crds-angular/App_Start/UnityConfig.cs
@@ -36,6 +36,7 @@ namespace crds_angular
             container.RegisterType<IContactRelationshipService, ContactRelationshipService>();
             container.RegisterType<IOpportunityService, OpportunityServiceImpl>();
             container.RegisterType<IConfigurationWrapper, ConfigurationWrapper>();
+            container.RegisterType<IServeService, ServeService>();
 
 
             container.RegisterType<PlatformServiceClient>(WithLifetime.PerResolve(typeof(PlatformServiceClient)), new InjectionConstructor());
