+++ Gateway/crds-angular/crds-angular.csproj
@@ -241,6 +241,7 @@
     <Compile Include="Models\Crossroads\CookieInfo.cs" />
     <Compile Include="Models\Crossroads\FamilyMember.cs" />
     <Compile Include="Models\Crossroads\ServeOccurance.cs" />
+    <Compile Include="Models\Crossroads\ServeResponseDto.cs" />
     <Compile Include="Models\Crossroads\SignUpFamilyMembers.cs" />
     <Compile Include="Models\Crossroads\GroupOpportunity.cs" />
     <Compile Include="Models\Crossroads\Serve\EventType.cs" />
